Scale grid columns only for full-window grids
Scaling the column count where LocalGridSettings is provided also
doubled it on surfaces that are capped at 640dp (edit favorites and
edit tag sheets, icon picker, hidden items sheet) and in settings
previews, cramming twice as many cells into a container that does not
grow with the window.

Provide the unscaled settings again and apply the window-width-class
scaling explicitly at the grids that actually span the window: the app
drawer/search results column and the search favorites grid. This also
removes a stale-capture hazard in ProvideSettings, where window size
and density were captured in a keyless remember block and only stayed
correct because a fold/unfold recreates the activity.
